refactor: 아키텍쳐 변경
보다 적절해보이는 위치로 변경

- utils 디렉토리 추가 후 validator, random, dto를 두도록 변경
- RacingCarManager를 model 디렉토리로 변경
- 위의 변화로 변경된 import 문에 변경
