+++ src/main/java/racingcar/manager/IOManager.java
@@ -1,10 +1,11 @@
 package racingcar.manager;
 
 import static racingcar.constant.InputStringConstant.INPUT_NAME_DIVIDER;
-import static racingcar.validate.InputValidator.validateName;
-import static racingcar.validate.InputValidator.validateRoundNum;
+import static racingcar.utils.validate.InputValidator.validateName;
+import static racingcar.utils.validate.InputValidator.validateRoundNum;
 
 import java.util.Arrays;
+import racingcar.utils.dto.Names;
 import racingcar.view.Input;
 import racingcar.view.Output;
