+++ src/main/java/racingcar/utils/dto/Names.java
@@ -1,4 +1,4 @@
-package racingcar.manager;
+package racingcar.utils.dto;
 
 import java.util.List;
