+++ src/main/java/racingcar/utils/validate/GameValidator.java
@@ -1,0 +1,4 @@
+package racingcar.utils.validate;
+
+public class GameValidator {
+}
