--- src/main/java/racingcar/validate/GameValidator.java
@@ -1,4 +1,0 @@
-package racingcar.validate;
-
-public class GameValidator {
-}
